# Remove sumargs, SIMD only for summation