This solves unexpected slow AD calculations. For example, `norm` becomes 4 times faster. I don't know why...

Before:

```jl
julia> using Tensorial, BenchmarkTools

julia> x = rand(SymmetricSecondOrderTensor{3});

julia> @benchmark gradient(norm, $(Ref(x))[], :all)
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  27.108 ns … 61.998 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     28.196 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   28.419 ns ±  1.415 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▁▄    █▃  ▂▃                                               ▁
  ▂██▄▄▇▇██▆▆███▆▄▆▄▅▄▄▅▃▃▂▅▄▃▄▂█▄▅▅▄▅▅▄▄▅▅▅▅▅▅▄▄▅▃▃▂▃▂▃▄▄▅▃▄ █
  27.1 ns      Histogram: log(frequency) by time      35.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

After:

```jl
julia> using Tensorial, BenchmarkTools

julia> x = rand(SymmetricSecondOrderTensor{3});

julia> @benchmark gradient(norm, $(Ref(x))[], :all)
BenchmarkTools.Trial: 10000 samples with 1000 evaluations.
 Range (min … max):  7.041 ns … 33.791 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     7.333 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.301 ns ±  0.590 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

             ▁     █                      █     █
  ▂▁▁▁▁▂▁▁▁▁▁█▁▁▁▁▁█▁▁▁▁▁▃▁▁▁▁▁▂▁▁▁▁▃▁▁▁▁▁█▁▁▁▁▁█▁▁▁▁▁▃▁▁▁▁▂ ▃
  7.04 ns        Histogram: frequency by time        7.46 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```